Emits reload to systemd when needed
The current implementation enforce the reload
when a 'wants' target is created or deleted.
This should work well for system units.
However, for system units, this behaviour isn't
enought when more that a user is active because
only the user that installs the application will
be updated. For this reason, a paralelle mechanism
has to be defined.

Also note that systemd is henceforth required for
tools because wgtpkg-installer needs it now.

Change-Id: I4fc03a44dbc58c2374ea21dbf6b436f646d04e00
Signed-off-by: José Bollo <jose.bollo@iot.bzh>