wgtpkg-digsig: Add flag to accept/refuse a WGT without signature
This flag allows with a single function to check the signature
and the case that a not signed widget is to be refused in the same
way that the signature is wrong.

The CMAKE boolean ALLOW_NO_SIGNATURE can be used to control
the default behavior at compile time. By default its value
is OFF meaning that install will refuse widgets without
signature.

Relates to Bug-AGL: SPEC-1590

Change-Id: I403109272759454696a1e5d9913879aaea7676e6
Signed-off-by: José Bollo <jose.bollo@iot.bzh>