Creates systemd service for APIs
Before this change, APIs were only provided as binder's
exported websockets. This forbade ability to use symbolic
links.

Allowing links is interesting for having platform sockets
and for linked api (not using socket).

The API afm-main is a platform API. It must run at system level
and must be available for all user clients. To achieve it, we
use a link.

Bug-AGL: SPEC-1592

Change-Id: I8753f83373755eb7fc2d2cc50251b8f738b61b03
Signed-off-by: José Bollo <jose.bollo@iot.bzh>