afm-urun: Force scan of systemd units
It was previously assumed that only launched units
had to be scanned. But it was buggy because the
uninstall/install of application rebuilds the database
without taking care of the running applications.

Nevertheless changing the behaviour of the rebuild
of application's database is not the best solution
for two reasons. The fist is that it will complicates
the code. The second, much better, is that it would
allow to track also applications launched automaticaly
-and/or manually in devel-.

Bug-AGL: SPEC-1836

Change-Id: I5fb2bb712ce43be9c851251be6b794a007b1ea6f
Signed-off-by: José Bollo <jose.bollo@iot.bzh>