Allow setting of libexec location
The location of the framework binding could not
be tuned but is fixed.

The default location is good but it is important
to be able to tune that location. This at least
has the good effect that the location is set at
one point in the build system instead of at multiple
points as today.

Bug-AGL: SPEC-2367

Change-Id: Ib06fdff56d1828eacd0d1f82c2e1308d0b36fa5e
Signed-off-by: José Bollo <jose.bollo@iot.bzh>