Refactor of system and user setup
Since a long time, the creation of the user and
system run environment was made by a sub-optimal,
hardly maintainable and false-error generating.

This change puts the user and system setup
in separate shell scripts. This makes them
more easy to maintain, update or tune.

This facility will useful for further evolution.

Bug-AGL: SPEC-1016

Change-Id: Iede81a659eacf4e5c21c561a33d300408e1bd058
Signed-off-by: Jose Bollo <jose.bollo@iot.bzh>