afm-unit.conf: Refactor its generation
The process of generating the file /etc/afm-unit.conf
had be reworked to be more straight forward.

Before that commit, a manual operation (a "make") had
to be done before committing the repository. This was
not efficient, error prone and kept temporary files in
conf.

That commit changes it by calling m4 through cmake.

Bug-AGL: SPEC-2436

Change-Id: Ia32a810286471dde8a01387d157e33277d67411d
Signed-off-by: Jose Bollo <jose.bollo@iot.bzh>