Add WAIT_FOR_HOST_SERVICE envvar to unit files for webapps
This envvar will make the application launcher wait for the WebAppMgr
service to be fully running before attempting to launch the app.
Otherwise, launcher would resort to running the app in standalone
mode, which would conflict with the WebAppMgr service when it's
finally running.

It will be useful for webapps that are auto-started, like the future
homescreen or the html5-launcher.

Bug-AGL: SPEC-2793

Signed-off-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>
Change-Id: I97f927c856b0dd2643f8b7492dff06b2c459761d