Add discovery of API
This change allows binders to use API not declared in
the config.xml file.

This behaviour is allowed by the platform permission

    urn:AGL:permission::platform:apis:auto-ws

It allows a binding to use started bindings not listed
in their config.xml file. But it does not allow to start
them or to discover them.

Bug-AGL: SPEC-2871

Change-Id: Iff0ec8417541171d446254e538201b7f6c7887ea
Signed-off-by: José Bollo <jose.bollo@iot.bzh>
(cherry picked from commit f68b681e785456207fd8195ac83ade0c89373416)