Add feature of autosigning widgets
This adds the ability to automatically sign the
widgets that are packaged.

This is done by defining in the environment of
the packaging process the variables

  WGTPKG_AUTOSIGN_X=key-filepath[:cert-filepath]...

Where X is a number.

If such variable exist, signatures are generated
in the directory of the packaged or signed widget,
one for each variable, replacing any existing one.

Obviously, nothing is done if no such variable exist.

The generated signature file depends on X.

 - 0 is for file author-signature.xml
 - X is for file signature-X.xml

The program wgtpkg-pack automatically include that
behaviour by default. An option allows to remove it.

Bug-AGL: SPEC-2840

Change-Id: I00bc4a4d094f71b307e467f984f20d3d4cc3c7bd
Signed-off-by: José Bollo <jose.bollo@iot.bzh>