Fix uninstall of widgets without icons
Trying to uninstall a widget without an icon would remove the widget
files, but report an error, and the widget would not be removed from
the database until a reboot, preventing installation of a new version.

To fix this, the error handling in uninstall_widget has been reworked
to only explicitly return an error when the unlink of the icon file
fails for reasons other than the file not being present. This then
allows some code cleanup by removing the extra rc2 variable and fixing
some error checks where rc and rc2 were being mixed.

Bug-AGL: SPEC-3401

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ia86805bd814e787b7e47c489973ee974fff4b6d3