Avoid RPM plugin failures
At the moment, when installing RPMs during image creation,
there is no way to contact any server and there is no fallback.

When creating a redpesk image that issue is only seen in recovery
creation. But in recovery, the security framework is not active
and then there is no issue to remove that feature.

In all cases, because there is no way to report a fatal error to
RPM in "post" callbacks, the real status of the transaction is to
be queried usin afmpkg-status.