Improve support of legacy config.xml
Add processing of 'provided-binding'.
Legacy installer was processing the provided-lib to make
public the provided libraries. This commit provides the
same behavior.

At the same time, it appeared that the below features
are no more processed by the framework:

- urn:AGL:widget:public-files
- urn:AGL:widget:lib-files
- urn:AGL:widget:conf-files
- urn:AGL:widget:exec-files
- urn:AGL:widget:data-files
- urn:AGL:widget:http-files

But the feature files-properties fit the same behavior and so
this undocumented features are now deprecated.
So for a smooth transition, the legacy file property settings are
automatically translated to the new version.