Ignore failure when work directory can't be created
This changes comes from working on integration with CGROUPS.
When working with CGROUPS, it might be needed to add the systemd
directive 'Delegate=yes' when CGROUP are monitored by applications,
as is the case of red-pak.

When Delegate=yes is active, the directive ExecStartPre is not
executed with delegation activated (man systemd.resource-control).
This lead redwrap to fail on CGROUP setting in ExecStartPre.

So it is assumed that rare applications launched with redwrap and
using CGROUP should manage the working directory themselves.
It can be seen on the previous line (WorkingDirectory=...) that
error of working directory are already ignored.